refactor: consolidate rust modules into a single module This change combines three separate Rust modules (ldk, rustbitcoin, and rustminiscript) into a single unified "rust" module. This simplifies the codebase by:
Removing duplicate build infrastructure across Rust modules
Creating a single entry point for all Rust-based functionality
Combining dependencies into a single Cargo.toml
Reducing the complexity of the main.cpp file

The changes include:

Deletion of individual module directories (ldk, rustbitcoin, rustminiscript)
Updated main.cpp to reference the new consolidated module
Modified the module loading code to use RustModules instead of individual modules

This refactoring should make the codebase easier to maintain while preserving
all existing functionality. And fix the issue of linking errors between
the Rust modules.
