+++ main.cpp
@@ -6,16 +6,8 @@
 #include <modules/bitcoin/module.h>
 #endif
 
-#ifdef RUST_BITCOIN
-#include <modules/rustbitcoin/module.h>
-#endif
-
 #ifdef MAKO
 #include <modules/mako/module.h>
-#endif
-
-#ifdef RUST_MINISCRIPT
-#include <modules/rustminiscript/module.h>
 #endif
 
 #ifdef BTCD
@@ -26,12 +18,12 @@
 #include <modules/lnd/module.h>
 #endif
 
-#ifdef LDK
-#include <modules/ldk/module.h>
-#endif
-
 #ifdef NLIGHTNING
 #include <modules/nlightning/module.h>
+#endif
+
+#ifdef RUST_MODULES
+#include <modules/rust/module.h>
 #endif
 
 std::shared_ptr<bitcoinfuzz::Driver> driver = nullptr;
@@ -43,14 +35,8 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t *Data, size_t Size)
 #ifdef BITCOIN_CORE
   driver->LoadModule(std::make_shared<bitcoinfuzz::module::Bitcoin>());
 #endif
-#ifdef RUST_BITCOIN
-  driver->LoadModule(std::make_shared<bitcoinfuzz::module::Rustbitcoin>());
-#endif
 #ifdef MAKO
   driver->LoadModule(std::make_shared<bitcoinfuzz::module::Mako>());
-#endif
-#ifdef RUST_MINISCRIPT
-  driver->LoadModule(std::make_shared<bitcoinfuzz::module::Rustminiscript>());
 #endif
 #ifdef BTCD
   driver->LoadModule(std::make_shared<bitcoinfuzz::module::Btcd>());
@@ -58,11 +44,11 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t *Data, size_t Size)
 #ifdef LND
   driver->LoadModule(std::make_shared<bitcoinfuzz::module::Lnd>());
 #endif
-#ifdef LDK
-  driver->LoadModule(std::make_shared<bitcoinfuzz::module::Ldk>());
-#endif
 #ifdef NLIGHTNING
   driver->LoadModule(std::make_shared<bitcoinfuzz::module::NLightning>());
+#endif
+#ifdef RUST_MODULES
+  driver->LoadModule(std::make_shared<bitcoinfuzz::module::RustModules>());
 #endif
 
   driver->Run(Data, Size, target);
