--- modules/ldk/ldk_lib/.gitignore
@@ -1,1 +1,0 @@
-target
