--- modules/ldk/ldk_lib/ldk_lib.h
@@ -1,3 +1,0 @@
-#include <cstdint>
-
-extern "C" bool ldk_des_invoice(const char* input);
