--- modules/rustbitcoin/rust_bitcoin_lib/.gitignore
@@ -1,1 +1,0 @@
-target
