--- modules/rustbitcoin/rust_bitcoin_lib/Cargo.toml
@@ -1,11 +1,0 @@
-[lib]
-name = "rust_bitcoin_lib"
-crate-type = ["staticlib"]
-
-[package]
-name = "rust_bitcoin_lib"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-bitcoin = "0.32.5"
