--- modules/rustminiscript/rust_miniscript_lib/Cargo.toml
@@ -1,11 +1,0 @@
-[lib]
-name = "rust_miniscript_lib"
-crate-type = ["staticlib"]
-
-[package]
-name = "rust_miniscript_lib"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-miniscript = { version = "12.3.0", features = ["std", "compiler"] }
