--- modules/rustminiscript/rust_miniscript_lib/rust_miniscript_lib.h
@@ -1,4 +1,0 @@
-#include <cstdint>
-
-extern "C" bool rust_miniscript_descriptor_parse(const char* input);
-extern "C" bool rust_miniscript_miniscript_parse(const char* input);
